requirements.txt
# flask
# flask-cors
# python-dotenv
# requests
# apscheduler
# playwright==1.41.1

Flask==3.0.3
Flask-Cors==4.0.0
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
APScheduler==3.10.4
playwright==1.49.0
gunicorn==22.0.0

